+++ eras/babbage/impl/CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add `getReferenceScriptsNonDistinct`
 * Add the constructor `BabbageNonDisjointRefInputs` to `BabbageUtxoPredFailure`
   * Utxo rule raises that `PredicateFailure` in Conway and future Eras when they are not disjoint.
+* Modify `PParams` JSON instances to match `cardano-api`
 
 ## 1.6.0.0