+++ eras/shelley/impl/CHANGELOG.md
@@ -21,6 +21,7 @@
 * Deprecate `RewardAcnt` in favor of `RewardAccount`
 * Remove `registerInitialFunds` and `registerInitialStaking`
 * Add `registerInitialFundsThenStaking`
+* Modify `PParams` JSON instances to match `cardano-api`
 
 ### `testlib`